+++ joulescope_ui/plugins/example_01.py
@@ -42,5 +42,5 @@ def plugin_register(api):
     :param api: The :class:`PluginServiceAPI` instance.
     :return: True on success any other value on failure.
     """
-    api.range_tool_register('Example 01', MyPlugin)
+    api.range_tool_register('Example/01', MyPlugin)
     return True