+++ joulescope_ui/plugins/example_02.py
@@ -40,5 +40,5 @@ def plugin_register(api):
     :param api: The :class:`PluginServiceAPI` instance.
     :return: True on success any other value on failure.
     """
-    api.range_tool_register('Example 02', MyPlugin)
+    api.range_tool_register('Example/02', MyPlugin)
     return True