+++ joulescope_ui/plugins/histogram.py
@@ -133,5 +133,5 @@ def plugin_register(api):
     :param api: The :class:`PluginServiceAPI` instance.
     :return: True on success any other value on failure.
     """
-    api.range_tool_register('Histogram', Histogram)
+    api.range_tool_register('Analysis/Histogram', Histogram)
     return True